+++ app/src-dao/com/geeknewbee/doraemon/database/PersonDao.java
@@ -23,7 +23,7 @@ public class PersonDao extends AbstractDao<Person, Long> {
      */
     public static class Properties {
         public final static Property Id = new Property(0, Long.class, "id", true, "_id");
-        public final static Property PersonId = new Property(1, Long.class, "personId", false, "PERSON_ID");
+        public final static Property PersonId = new Property(1, Integer.class, "personId", false, "PERSON_ID");
         public final static Property Name = new Property(2, String.class, "name", false, "NAME");
     }
 
@@ -45,38 +45,34 @@ public class PersonDao extends AbstractDao<Person, Long> {
      * Creates the underlying database table.
      */
     public static void createTable(SQLiteDatabase db, boolean ifNotExists) {
-        String constraint = ifNotExists ? "IF NOT EXISTS " : "";
+        String constraint = ifNotExists ? "IF NOT EXISTS ": "";
         db.execSQL("CREATE TABLE " + constraint + "\"PERSON\" (" + //
                 "\"_id\" INTEGER PRIMARY KEY AUTOINCREMENT ," + // 0: id
                 "\"PERSON_ID\" INTEGER," + // 1: personId
                 "\"NAME\" TEXT);"); // 2: name
     }
 
-    /**
-     * Drops the underlying database table.
-     */
+    /** Drops the underlying database table. */
     public static void dropTable(SQLiteDatabase db, boolean ifExists) {
         String sql = "DROP TABLE " + (ifExists ? "IF EXISTS " : "") + "\"PERSON\"";
         db.execSQL(sql);
     }
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     @Override
     protected void bindValues(SQLiteStatement stmt, Person entity) {
         stmt.clearBindings();
-
+ 
         Long id = entity.getId();
         if (id != null) {
             stmt.bindLong(1, id);
         }
 
-        Long personId = entity.getPersonId();
+        Integer personId = entity.getPersonId();
         if (personId != null) {
             stmt.bindLong(2, personId);
         }
-
+ 
         String name = entity.getName();
         if (name != null) {
             stmt.bindString(3, name);
@@ -89,49 +85,39 @@ public class PersonDao extends AbstractDao<Person, Long> {
         entity.__setDaoSession(daoSession);
     }
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     @Override
     public Long readKey(Cursor cursor, int offset) {
         return cursor.isNull(offset + 0) ? null : cursor.getLong(offset + 0);
     }
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     @Override
     public Person readEntity(Cursor cursor, int offset) {
         Person entity = new Person( //
                 cursor.isNull(offset + 0) ? null : cursor.getLong(offset + 0), // id
-                cursor.isNull(offset + 1) ? null : cursor.getLong(offset + 1), // personId
+                cursor.isNull(offset + 1) ? null : cursor.getInt(offset + 1), // personId
                 cursor.isNull(offset + 2) ? null : cursor.getString(offset + 2) // name
         );
         return entity;
     }
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     @Override
     public void readEntity(Cursor cursor, Person entity, int offset) {
         entity.setId(cursor.isNull(offset + 0) ? null : cursor.getLong(offset + 0));
-        entity.setPersonId(cursor.isNull(offset + 1) ? null : cursor.getLong(offset + 1));
+        entity.setPersonId(cursor.isNull(offset + 1) ? null : cursor.getInt(offset + 1));
         entity.setName(cursor.isNull(offset + 2) ? null : cursor.getString(offset + 2));
     }
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     @Override
     protected Long updateKeyAfterInsert(Person entity, long rowId) {
         entity.setId(rowId);
         return rowId;
     }
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     @Override
     public Long getKey(Person entity) {
         if (entity != null) {
@@ -144,9 +130,9 @@ public class PersonDao extends AbstractDao<Person, Long> {
     /**
      * @inheritdoc
      */
-    @Override
+    @Override    
     protected boolean isEntityUpdateable() {
         return true;
     }
-
+    
 }
